2007-07-23 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* src/components/ogre/terrain/TerrainArea.cpp, src/components/ogre/terrain/TerrainGenerator.cpp: use size_t where suitable

+++ src/components/ogre/terrain/TerrainArea.cpp
@@ -74,7 +74,7 @@ bool TerrainArea::parseArea()
     }
        
     WFMath::Polygon<2> poly;
-    for (unsigned int p=0; p<pointsData.size(); ++p) {
+    for (size_t p=0; p<pointsData.size(); ++p) {
         if (!pointsData[p].isList()) {
             S_LOG_FAILURE("skipped malformed point in area");
             continue;
